Use first letter of tracer name as coordinate name
So far, the coordinates of the Datasets returned by the sort functions
were called `var_q/Q`, which does not convey any meaning.  With this
change, the coordinates are named after the first letter of the tracer
that was used for binning, e.g., `s/S` for salinity bins.  There are 2
exceptions to this rule:  (1) If a tracer name starts with q, then the
coordinate is named `c` instead, because `q` is the name of the
transport.  (2) In the case of 2-dimensional sorting, if both
coordinates would end up with the same name, the second coordinate name
is extended by `2`, like previously.

Furthermore, the attributes of the tracers that were used for binning,
e.g., long name and units, are copied to the new coordinates.  This
gives more sensible axes labels when plotting the data with xarray.