Fix the indices for binning
The indicies given by np.digitize are 1 larger than expected, which had
the consequence that the bins were always off by 1.  This is now fixed.

Using xarray's apply_ufunc is actually not needed for np.digitize, since
the code following it also works for numpy arrays, so it is removed now.