Add option to bin volume and tracer transport at once
A new option is added to sort_1dim that allows to sort volume transport
and tracer transport into bins at once and include both in one Dataset.
Usually, the same bins are used for both transports.  In the frequent
case that the tracer transport is simply the product of tracer and
transport, this can be computed in the sort_1dim function, so the user
has to call the function only once instead of twice, ends up with one
output Dataset instead of two, and does not have to change the value of
tef.transport.  This can simplify user code a lot and does not break
any existing code.