Add option to get volume and tracer bulk values at once
Previously, it was necessary to calculate the bulk values for volume
transport and tracer transport separately.  It was possible to provide
the indices of the bulk values of the first calculation to the second,
which reduced the number of computations.  However, with this method,
the information got lost which bulk value corresponds to inflow and
which to outflow.  This can be a problem when the tracer has both
positive and negative values, like temperature in high latitudes.
By computing both bulk values at once, even less computations are needed
and it is clear which one corresponds to in- and which to outflow.
Also, it is easier for the user, since only a single Dataset is returned
instead of two separate Datasets with results.  This option is
implemented with this commit.  As it is optional, it does not break any
existing code.